% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados_exemplo.R
\docType{data}
\name{dados_exemplo}
\alias{dados_exemplo}
\title{Dados de exemplo}
\format{
`data frame`
Um dataframe contendo duas colunas
\describe{
  \item{col1}{Numeros de 1 a 10}
  \item{col2}{Numeros de 10 a 1}
  ...
}
}
\usage{
dados_exemplo
}
\description{
Duas colunas de numeros inteiros de 1 a 10
}
\keyword{datasets}
