% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Soma.R
\name{Soma}
\alias{Soma}
\title{Soma}
\usage{
Soma(x, y)
}
\arguments{
\item{x}{Um numero.}

\item{y}{Um numero.}
}
\value{
A soma dos dois numeros.
}
\description{
Soma dois numeros
}
\examples{
Soma(1, 1)
Soma(10, 1)
}
