% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomaNumeros.R
\name{SomaNumeros}
\alias{SomaNumeros}
\title{SomaNumeros}
\usage{
SomaNumeros(x, y)
}
\arguments{
\item{x}{Um numero.}

\item{y}{Outro Numero}
}
\value{
A soma dos dois numeros e faz um gráfico
}
\description{
Soma dois numeros
}
